ADC modification
- adc kranked to 512 MHz, internal 8 MHz -> measured power consumption is still 40 mA (not changed)
- Modified .ioc files
- ADC auto calibration implemented in measure.c
- Temperature sensor HAL driver include, temp sensor is working now

+++ Core/Inc/Measurements.h
@@ -13,7 +13,7 @@
 
 void measure(Request request);
 uint16_t analogRead();
-uint16_t get_temperature();
+int get_temperature();
 void select_measure_adc();
